Added a newert datetimepicker with linked pickers Added a schedule of the bookings in admin... Google rank tag added th:switch to car.transmission to display in bg Added another service to services and updated grid layout Added meta site description Added LPG instead of TRAFFIC in availables page Fixed carousel Fixed paypal accounts Added more info in banner

+++ src/main/kotlin/com/rentautosofia/rentacar/service/PaypalService.kt
@@ -22,7 +22,7 @@ class PaypalService {
 
     @Throws(PayPalRESTException::class)
     fun createPayment(
-            total: Double?,
+            total: Float?,
             currency: String,
             method: PaypalPaymentMethod,
             intent: PaypalPaymentIntent,
